requirements.txt
Django~=4.2.7
django-extensions~=3.2.3
django-tinymce~=3.6.1
django-admin-sortable2~=2.1.10
Pillow~=10.1.0

#проверка кода
flake8~=6.1.0
isort~=5.12.0
